hange mainPage component to LoginedMain and Main has logined,notlogined page
