+++ src/Pages/Login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { Box, Image, Stack } from "@chakra-ui/react";
+import { Stack } from "@chakra-ui/react";
 
 import LoginForm from "../Components/LoginForm/LoginFom";
 import Footer from "../Components/Footer/Footer.js";
